% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data matrix}
\usage{
normalize_data(data)
}
\arguments{
\item{data}{Numeric matrix (rows = variables, columns = samples).}
}
\value{
Normalized data matrix, same dimensions as \code{data}.
}
\description{
Normalize data with respect to the sample (column) total.
}
