% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_data.R
\name{shuffle_data}
\alias{shuffle_data}
\title{Column-wise shuffle the data matrix}
\usage{
shuffle_data(data)
}
\arguments{
\item{data}{Data matrix (rows are variables, columns are samples).}
}
\value{
Shuffled data matrix, same dimensions as \code{data}.
}
\description{
Shuffle each column of the data matrix to remove any correlations between
variables but preserve the sample-specific profiles.
}
\concept{expressiondata}
