% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity_network.R
\name{sparsify_network}
\alias{sparsify_network}
\title{Remove indirect links from the network using network sparsification}
\usage{
sparsify_network(mi)
}
\arguments{
\item{mi}{Network (MI) matrix. Assumed symmetric.}
}
\value{
A list with:
\item{sps}{Sparsified network matrix.}
\item{gcc}{Global clustering coefficient.}
}
\description{
Removes links from the network that are the result of indirect
correlations, via sparsification based on the Data Processing
Inequality. Iterates over all triangles in the network and removes the
weakest link in each. Also used to calculate the network Global
Clustering Coefficient (GCC).
}
